version 7.0.4 working dev center for vsphere

+++ local/build.sh
@@ -6,7 +6,7 @@
 #
 # This script builds an web server image in the file system of the VMware {code}
 # "Devcenter App" https://gitlab.eng.vmware.com/gtix-tools/vcode-dev-center-app
-# 
+#
 # Read https://gitlab.eng.vmware.com/gtix-tools/vcode-dev-center-app/blob/master/INTEGRATION.md'
 # for instructions on integration.
 #
@@ -18,14 +18,18 @@ SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 # -----------------------------------------------------------------------------
 # VARIABLES YOU CAN SET
 
-# you can supply an overridden build directory via BUILD_DIR variable if you 
+# you can supply an overridden build directory via BUILD_DIR variable if you
 # wish.  If not provided, all output is created in this (the script) directory.
 BUILD_DIR=${BUILD_DIR:-${SCRIPT_DIR}}
 
-# the VER variable is the one place to change the particular release of API 
+# if your dev-center is not in the server root, you must supply is not in the root (likely)
+# you must supply this path here by setting DEV_CENTER_ROOT_PATH
+DEV_CENTER_ROOT_PATH=${DEV_CENTER_ROOT_PATH:-""}
+
+# the VER variable is the one place to change the particular release of API
 # explorer.  See https://build-artifactory.eng.vmware.com/artifactory/npm/%40vmw/vcode-dev-center-app/-/@vmw
 # for valid values
-export VER="7.0.1"
+export VER="7.0.4"
 
 # -----------------------------------------------------------------------------
 ARTIFACTORY_URL=https://build-artifactory.eng.vmware.com/artifactory/npm/%40vmw/vcode-dev-center-app/-/@vmw
@@ -41,6 +45,12 @@ mkdir -p ${DOWNLOAD_DIR}
 if [ ! -f ${DOWNLOAD_DIR}/${DEV_CENTER_APP_FILE} ]; then
     echo "Downloading archives of the distribution to ${DOWNLOAD_DIR}/${DEV_CENTER_APP_FILE}"
     wget --no-check-certificate ${ARTIFACTORY_URL}/${DEV_CENTER_APP_FILE} --output-document ${DOWNLOAD_DIR}/${DEV_CENTER_APP_FILE}
+
+    if [ $? -ne 0 ]; then
+        echo "Failed to download dev-center-app from ${ARTIFACTORY_URL}/${DEV_CENTER_APP_FILE}"
+        exit 1
+    fi
+
     pushd ${DOWNLOAD_DIR}
     tar xfz ${DEV_CENTER_APP_FILE}  #this creates ${DOWNLOAD_DIR}/package/dist and ${DOWNLOAD_DIR}/package/tools
     rm -rf ${TOOLS_DIR}  # if we downloaded new tools, wipe the old ones
@@ -64,18 +74,17 @@ rm -rf ${OUTPUT_DIR}/*
 mkdir -p ${OUTPUT_DIR}
 pushd ${OUTPUT_DIR}
 
-echo "Extracting dev-center distribution"
-cp -R ${DOWNLOAD_DIR}/package/dist/.* .
-cp -R ${DOWNLOAD_DIR}/package/dist/* .
+echo "Staging dev-center distribution"
+cp -R ${DOWNLOAD_DIR}/package/dist/{.[^.],}* .
 
 echo "Cleaning up unneeded files in distribution"
 rm -fv ./assets/apix-swagger.json
 rm -fv ./assets/dev-center-overview.html
 rm -fv ./assets/sample-exchange-swagger.json
 #rm -fv ./assets/swagger-auth.json
-#rm -fv ./assets/swagger-complete.json
+rm -fv ./assets/swagger-complete.json
 rm -fv ./assets/vra-auth.json
-#rm -fv ./assets/vra-config.json
+rm -fv ./assets/vra-config.json
 rm -rfv ./local
 
 echo "Removing stock configuration files (will be overwritten, but to catch errors we remove them explicitly)"
@@ -86,34 +95,23 @@ echo "Staging product specific dev-center-config.json as well as apis.json"
 cp -fv ${SCRIPT_DIR}/dev-center-config.json ./assets
 cp -fv ${SCRIPT_DIR}/apis.json ./assets
 cp -fv ${SCRIPT_DIR}/favicon.ico ./    # copy VMware standard icon.  Replace with your products if there is one.
+cp -fv ${SCRIPT_DIR}/dev-center-overview.html ./
+
+echo "Copying swagger files"
+mkdir -p ./assets/swagger
+cp -fv ${SCRIPT_DIR}/swagger/* ./assets/swagger
+
+echo "Using DEV_CENTER_ROOT_PATH variable to set base path in generated files to \"${DEV_CENTER_ROOT_PATH}\""
 
 
-# The below commented code is an attempt to create a static war file wrapper for the content
-#WAR_OUTPUT_DIR=${BUILD_DIR}/staging-war
-#rm -rf ${WAR_OUTPUT_DIR}/*
-#mkdir -p ${WAR_OUTPUT_DIR}
-#WAR_FILE_NAME=dev-center.war
-#pushd ${WAR_OUTPUT_DIR}
-#echo "Staging war metadata"
-#cp -Rv ${SCRIPT_DIR}/war-template/* ${WAR_OUTPUT_DIR}
-#mkdir -p ${WAR_OUTPUT_DIR}/web
-#cp -Rv ${OUTPUT_DIR}/* ${WAR_OUTPUT_DIR}/web
-#echo "Creating war file ${PUBLISH_DIR}/${WAR_FILE_NAME}"
-#zip -r ${PUBLISH_DIR}/${WAR_FILE_NAME} *
-#popd
-#echo "Creating war metadata"
-# now create a war file that is simply a wrapper on the image 
+sed -i "s|<base href=\\\"/\\\">|<base href=\\\"${DEV_CENTER_ROOT_PATH}\\\">|g" ./index.html
+sed -i "s|VMware {code} Developer Center App|vSphere Developer Center|g" ./index.html
+sed -i "s|/dev-center-app/|/${DEV_CENTER_ROOT_PATH}/|g" ./.htaccess
+sed -i "s|/dev-center-app/|/${DEV_CENTER_ROOT_PATH}/|g" ./server.rewrites
 
-#mkdir -p ${OUTPUT_DIR}/WEB-INF
-#cat > ${OUTPUT_DIR}/WEB-INF/web.xml <<- "EOF"
-#<web-app version="3.0" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd">
-#<display-name>VMware {code} Services Dev Center</display-name>
-#<welcome-file-list><welcome-file>index.html</welcome-file></welcome-file-list>
-#</web-app>
-#EOF
-#echo "Creating war file ${PUBLISH_DIR}/${WAR_FILE_NAME}"
-#rm -f ${PUBLISH_DIR}/${WAR_FILE_NAME}
-#zip -r ${PUBLISH_DIR}/${WAR_FILE_NAME} *
+echo "Publishing output zip file ${PUBLISH_DIR}/dev-center.zip"
+mkdir -p ${PUBLISH_DIR}
+zip -r ${PUBLISH_DIR}/dev-center.zip {.[^.],}*
 
 popd
 echo "DONE!"
